Feat: 'admin' namespace for tests, BaseController
