+++ app/views/admin/tests/show.html.erb
@@ -4,5 +4,6 @@
 </p>
 <%= link_to 'New question', new_test_question_path(@test) %>
 <ol>
-  <%= render @test.questions %>
+  <%= render partial: 'questions/question', 
+  collection: @test.questions %>
 </ol>
