+++ app/views/application/_nav.html.erb
@@ -1,8 +1,13 @@
 <div class="container">
-  <% if user_signed_in? %>
   <p class="nav user">
+  <% if user_signed_in? %>
+  
     welcome, <b><%= current_user.email %></b> |
     <%= link_to 'Signout', destroy_user_session_path, data: { turbo_method: :delete } %>
-  </p>
+  <% else %>
+   
+    <%= link_to 'Signin', new_user_session_path %>   
+  
   <% end %>
+  </p>
 </div>
