+++ app/views/questions/show.html.erb
@@ -5,5 +5,5 @@
 <ol>
   <%= render @question.answers %>
 </ol>
-<%= link_to 'Edit', edit_question_path(@question) %>
-<%= link_to 'Back', test_path(@question.test) %>
+<%= link_to 'Edit', edit_admin_question_path(@question) %>
+<%= link_to 'Back', admin_test_path(@question.test) %>
