+++ app/views/tests/_test.html.erb
@@ -1,17 +1,8 @@
 <tr>
-  <td><%= test.author.name %> </td>
+  <td><%= test.title %></td>
   <td><%= test.category.title %></td>
   <td><%= test_level(test.level) %></td>
-  <td><%= link_to test.title, test_path(test) %></td>
   <td><%= test.questions.count %></td>
-  <td><%= link_to 'Править', edit_test_path(test) %>
-  </td>
-  <td>
-    <%= link_to 'Удалить', test_path(test), data: {
-                    turbo_method: :delete,
-                    turbo_confirm: 'Are you sure?'
-                  } %>
-  </td>
   <td><%= button_to 'Start', start_test_path(test),
     data: { turbo_method: :post } %></td>
 </tr>
