+++ app/views/tests/index.html.erb
@@ -1,16 +1,12 @@
 Список тестов
 <table>
   <tr>
-    <th>Автор</th>
-    <th>Категория</th>
-    <th>Уровень</th>
     <th>Тест</th>
+    <th>Категория</th>
+    <th>Уровень</th>    
     <th>Количество вопросов</th>
-    <th>Править</th>
-    <th>Удалить</th>
     <th>Прохождение</th>
   </tr>
   <%= render @tests %>
 </table>
 <br>
-<%= link_to 'Создать тест', new_test_path %>
